libpinproc: fix PRDevice::GetVersionInfo() API
`combinedVersionRevision` is a 32-bit value.  This API isn't exposed in
the public pinproc.h header, so it's safe to make this change.

+++ src/PRDevice.cpp
@@ -1383,7 +1383,7 @@ int PRDevice::CalcCombinedVerRevision()
     return 0;
 }
 
-int PRDevice::GetVersionInfo(uint16_t *verPtr, uint16_t *revPtr, uint16_t *combinedPtr)
+int PRDevice::GetVersionInfo(uint16_t *verPtr, uint16_t *revPtr, uint32_t *combinedPtr)
 {
     *verPtr = version;
     *revPtr = revision;
